fixed issue with not being able to make maps over a 255x255 blocks

+++ map.c
@@ -18,9 +18,9 @@ int random_walk(map_t *map,int width,int height,double walk_percent,int seed)
     }
 
     // init the map to all values of green gray scale
-    for(uint8_t x=0;x<MAP_WIDTH;x++)
+    for(uint32_t x=0;x<MAP_WIDTH;x++)
     {
-        for(uint8_t y=0;y<MAP_HEIGHT;y++)
+        for(uint32_t y=0;y<MAP_HEIGHT;y++)
         {
             uint8_t green_scale = (rand() % 256) + 1;
             RGB_t block = {0x00,green_scale,0x00};
